requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
pymssql
PyMySQL
python-dotenv
SQLAlchemy
XlsxWriter
bcrypt
